as: added some fixes and refactor
